[refactor] useUserState 훅을 만들어 app.tsx를 정리함
app.tsx 를 간결하게
